Add Format Data Function
I noticed you performed this action about three times and as it was a little long, I thought it might be more efficient if there was simply a defined function to perform this somewhat repetitive action. I may be a little off with regards to the structuring of the variables (e.g., global vs local variable usage), so please forgive me, but I believe you can see the general idea that I am proposing here. I also made these changes using the GitHub online vscode editor (by pressing dot when viewing this file on this website), but I had some issues and I did my best to propose my changes here as I previously attempted in the vscode editor (sorry if there are issues from it).